+++ pvinspect/common/exceptions.py
@@ -1,0 +1,6 @@
+class MisconfigurationException(Exception):
+    pass
+
+
+class InvalidArgumentException(Exception):
+    pass