+++ pvinspect/common/types.py
@@ -1,0 +1,6 @@
+from pathlib import Path
+from typing import Union, Dict, List, Tuple
+from shapely.geometry import Polygon
+
+PathOrStr = Union[Path, str]
+ObjectAnnotations = Dict[str, List[Tuple[str, Polygon]]]