+++ test/unit/test_preprocessing.py
@@ -1,13 +1,43 @@
 from pvinspect.data.image import *
 from pvinspect.preproc.calibration import *
-from pvinspect.preproc.process import *
-from pvinspect.preproc.calibration import _calibrate_flatfield
-from pvinspect.preproc.process import _compensate_flatfield
+from pvinspect.preproc.calibration import _calibrate_flatfield, _compensate_flatfield
 from pvinspect.data import datasets
 import numpy as np
 from skimage.exposure import rescale_intensity
 from test.utilities import assert_equal
 import cv2
+
+
+def _prepare_ff_data():
+    ds = datasets.calibration_ipv40CCD_FF(N=2)
+    del ds["1A"]
+    imgs_list = []
+    ex_list = []
+    ex_all = [int(x[0]) for x in ds.keys()]
+    ex_max = max(ex_all)
+    for ex in ex_all:
+        for img in ds["{:d}A".format(ex)]:
+            imgs_list.append(img)
+            ex_list.append(ex / ex_max)
+
+    return ImageSequence(imgs_list, same_camera=True), ex_list
+
+
+def _prepare_ff_test_img():
+    ds = datasets.calibration_ipv40CCD_FF(N=2)
+    del ds["1A"]
+    imgs_list = []
+    ex_list = []
+    ex_all = [int(x[0]) for x in ds.keys()]
+    ex_max = max(ex_all)
+    for ex in ex_all:
+        for img in ds["{:d}A".format(ex)]:
+            imgs_list.append(img)
+            ex_list.append(ex / ex_max)
+    test_img = ModuleImage(
+        ds["{:d}A".format(ex_all[1])][0]._data, EL_IMAGE, path=Path() / "test.png"
+    )
+    return test_img
 
 
 def _make_image_seq(imgs):
@@ -20,7 +50,7 @@ def test_calibrate_flatfield_linear():
     img1 = (-np.random.random((10, 10)) / 10) + 1.0
     coeff = 1 / (img1 - img0)
 
-    coeff_result = _calibrate_flatfield([img0, img1], [0, 1.0])
+    coeff_result = _calibrate_flatfield([img0, img1], [0, 1.0], order=1)
     assert_equal(coeff_result[1], coeff)
     assert_equal(coeff_result[0], -coeff * img0)
 
@@ -33,7 +63,7 @@ def test_calibrate_flatfield_quadratic():
     def compensate(img, coeff):
         return coeff[2] * img ** 2 + coeff[1] * img + coeff[0]
 
-    coeff = _calibrate_flatfield([img0, img1, img2], [0, 0.5, 1.0])
+    coeff = _calibrate_flatfield([img0, img1, img2], [0, 0.5, 1.0], order=2)
     assert_equal(np.zeros_like(img0), compensate(img0, coeff))
     assert_equal(np.full_like(img1, 0.5), compensate(img1, coeff))
     assert_equal(np.full_like(img1, 1.0), compensate(img2, coeff))
@@ -64,7 +94,7 @@ def test_calibrate_and_compensate_flatfield():
     img1 = np.random.random((10, 10)) / 10 + 0.5 - 0.05
     img2 = (-np.random.random((10, 10)) / 10) + 1.0
 
-    coeff = _calibrate_flatfield([img0, img1, img2], [0, 0.5, 1.0])
+    coeff = _calibrate_flatfield([img0, img1, img2], [0, 0.5, 1.0], order=2)
     compensated = _compensate_flatfield(_make_image_seq([img0, img1, img2]), coeff)
     assert_equal(np.zeros_like(img0), compensated[0].data)
     assert_equal(np.full_like(img1, 0.5), compensated[1].data)
@@ -82,20 +112,9 @@ def test_calibrate_and_compensate_preserves_range():
 
 
 def test_calibrate_and_compensate_flatfield_realdata():
-    ds = datasets.calibration_ipv40CCD_FF(N=2)
-    del ds["1A"]
-    imgs_list = []
-    ex_list = []
-    ex_all = [int(x[0]) for x in ds.keys()]
-    ex_max = max(ex_all)
-    for ex in ex_all:
-        for img in ds["{:d}A".format(ex)]:
-            imgs_list.append(img)
-            ex_list.append(ex / ex_max)
-    coeff = calibrate_flatfield(ImageSequence(imgs_list, same_camera=True), ex_list)
-    test_img = ModuleImage(
-        ds["{:d}A".format(ex_all[1])][0]._data, EL_IMAGE, path=Path() / "test.png"
-    )
+    imgs_list, ex_list = _prepare_ff_data()
+    test_img = _prepare_ff_test_img()
+    coeff = calibrate_flatfield(imgs_list, ex_list)
     compensated = compensate_flatfield(test_img, coeff)
     data = compensated.data
     assert data.min() >= 0.0
@@ -132,3 +151,27 @@ def test_calibrate_compensate_distortion_realdata():
 
     for img, comp in zip(ds, ds_comp):
         assert_equal(comp.data.mean(), img.data.mean(), 0.01 * img.data.max())
+
+
+def test_calibration_object(tmp_path: Path):
+    ff_imgs, ex_list = _prepare_ff_data()
+    test_img = _prepare_ff_test_img()
+    dist_imgs = datasets.calibration_ipv40CCD_distortion(N=8)
+
+    # perform calibration
+    calib = Calibration()
+    calib.calibrate_flatfield(images=ff_imgs, targets=ex_list)
+    calib.calibrate_distortion(images=dist_imgs, checkerboard_size=(7, 7))
+
+    # save to disk
+    calib.save(tmp_path / "calib.pck")
+
+    # load from disk
+    calib_loaded = load_calibration(tmp_path / "calib.pck")
+
+    # perform compensation with both
+    comp = calib.process(test_img)
+    comp_loaded = calib_loaded.process(test_img)
+
+    # assert both result in the same image
+    assert_equal(comp.data, comp_loaded.data)